shared: update constants and types for payments/nasab changes
