+++ shared/package.json
@@ -2,17 +2,18 @@
   "name": "@zakapp/shared",
   "version": "1.0.0",
   "description": "Shared types and utilities for zakapp",
+  "type": "module",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "require": "./dist/index.js",
       "import": "./dist/index.js",
+      "require": "./dist/index.js",
       "types": "./dist/index.d.ts"
     },
     "./types/tracking": {
-      "require": "./dist/types/tracking.js",
       "import": "./dist/types/tracking.js",
+      "require": "./dist/types/tracking.js",
       "types": "./dist/types/tracking.d.ts"
     }
   },
@@ -38,7 +39,7 @@
     "eslint": "^8.52.0",
     "jest": "^29.7.0",
     "prettier": "^3.0.3",
-    "typescript": "^5.2.2"
+    "typescript": "^5.9.3"
   },
   "files": [
     "dist"
