+++ shared/src/constants.ts
@@ -2,7 +2,7 @@
 
 /**
  * Re-export comprehensive Islamic constants from dedicated module
- * 
+ *
  * IMPORTANT: The canonical source for Islamic Zakat constants is now
  * in constants/islamicConstants.ts. This includes:
  * - NISAB_THRESHOLDS (with scholarly references)
@@ -10,26 +10,26 @@
  * - HAWL_CONSTANTS (lunar calendar calculations)
  * - CALCULATION_METHODS (madhab-specific approaches)
  * - Helper functions for calculations
- * 
+ *
  * The legacy constants below are maintained for backward compatibility
  * but new code should use the comprehensive islamicConstants module.
  */
-export * from './constants/islamicConstants';
+export * from './constants/islamicConstants.js';
 
 // Legacy Zakat calculation constants (DEPRECATED - use islamicConstants instead)
 // @deprecated Use ZAKAT_RATES from islamicConstants.ts
-export const ZAKAT_RATES = {
-  STANDARD_RATE: 2.5, // 2.5% for most assets
-  AGRICULTURAL_RAIN_FED: 10, // 10% for rain-fed crops
-  AGRICULTURAL_IRRIGATED: 5, // 5% for irrigated crops
-} as const;
+// export const ZAKAT_RATES = {
+//   STANDARD_RATE: 2.5, // 2.5% for most assets
+//   AGRICULTURAL_RAIN_FED: 10, // 10% for rain-fed crops
+//   AGRICULTURAL_IRRIGATED: 5, // 5% for irrigated crops
+// } as const;
 
 // Legacy Nisab thresholds (DEPRECATED - use islamicConstants instead)
 // @deprecated Use NISAB_THRESHOLDS from islamicConstants.ts
-export const NISAB_THRESHOLDS = {
-  GOLD_GRAMS: 87.48, // ~3 ounces of gold
-  SILVER_GRAMS: 612.36, // ~21 ounces of silver
-} as const;
+// export const NISAB_THRESHOLDS = {
+//   GOLD_GRAMS: 87.48, // ~3 ounces of gold
+//   SILVER_GRAMS: 612.36, // ~21 ounces of silver
+// } as const;
 
 // Supported currencies
 export const CURRENCIES = [
@@ -48,728 +48,7 @@ export const CURRENCIES = [
 ] as const;
 
 // Asset categories with their details
-export const ASSET_CATEGORIES = {
-  CASH: {
-    id: 'cash',
-    name: 'Cash & Bank Accounts',
-    description: 'Liquid cash, savings, and checking accounts',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'savings',
-        name: 'Savings Account',
-        description: 'Money in savings accounts',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['interestRate'],
-      },
-      {
-        id: 'checking',
-        name: 'Checking Account',
-        description: 'Money in checking accounts',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: [],
-      },
-      {
-        id: 'cash_on_hand',
-        name: 'Cash on Hand',
-        description: 'Physical cash in possession',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: [],
-      },
-      {
-        id: 'certificates_of_deposit',
-        name: 'Certificates of Deposit',
-        description: 'Time deposits with fixed terms',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['interestRate', 'maturityDate'],
-      },
-      {
-        id: 'money_market',
-        name: 'Money Market Accounts',
-        description: 'High-yield savings with limited transactions',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['interestRate'],
-      },
-    ],
-  },
-  GOLD: {
-    id: 'gold',
-    name: 'Gold',
-    description: 'Gold jewelry, coins, and bars',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'jewelry',
-        name: 'Gold Jewelry',
-        description: 'Gold jewelry for personal use or investment',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'coins',
-        name: 'Gold Coins',
-        description: 'Gold coins for investment',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'bars',
-        name: 'Gold Bars',
-        description: 'Gold bars and bullion',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'ornaments',
-        name: 'Gold Ornaments',
-        description: 'Decorative gold items',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-    ],
-  },
-  SILVER: {
-    id: 'silver',
-    name: 'Silver',
-    description: 'Silver jewelry, coins, and bars',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'jewelry',
-        name: 'Silver Jewelry',
-        description: 'Silver jewelry for personal use or investment',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'coins',
-        name: 'Silver Coins',
-        description: 'Silver coins for investment',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'bars',
-        name: 'Silver Bars',
-        description: 'Silver bars and bullion',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'ornaments',
-        name: 'Silver Ornaments',
-        description: 'Decorative silver items',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-      {
-        id: 'utensils',
-        name: 'Silver Utensils',
-        description: 'Silver utensils and household items',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['weight', 'purity'],
-      },
-    ],
-  },
-  BUSINESS: {
-    id: 'business',
-    name: 'Business Assets',
-    description: 'Business inventory, trade goods, and commercial assets',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'inventory',
-        name: 'Business Inventory',
-        description: 'Goods held for sale in ordinary course of business',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['businessType', 'holdingPeriod'],
-      },
-      {
-        id: 'trade_goods',
-        name: 'Trade Goods',
-        description: 'Goods acquired for trading purposes',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['businessType', 'holdingPeriod'],
-      },
-      {
-        id: 'raw_materials',
-        name: 'Raw Materials',
-        description: 'Materials to be used in production',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['businessType'],
-      },
-      {
-        id: 'finished_goods',
-        name: 'Finished Goods',
-        description: 'Completed products ready for sale',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['businessType'],
-      },
-      {
-        id: 'work_in_progress',
-        name: 'Work in Progress',
-        description: 'Partially completed goods',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['businessType'],
-      },
-    ],
-  },
-  PROPERTY: {
-    id: 'property',
-    name: 'Investment Property',
-    description: 'Real estate held for investment purposes',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'residential_investment',
-        name: 'Residential Investment',
-        description:
-          'Residential properties held for rental income or appreciation',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['propertyType', 'location', 'rentalIncome'],
-      },
-      {
-        id: 'commercial',
-        name: 'Commercial Property',
-        description: 'Office buildings, retail spaces, warehouses',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['propertyType', 'location', 'rentalIncome'],
-      },
-      {
-        id: 'land',
-        name: 'Land',
-        description: 'Undeveloped land held for investment',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['location'],
-      },
-      {
-        id: 'agricultural',
-        name: 'Agricultural Property',
-        description: 'Farm land and agricultural properties',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['propertyType', 'location'],
-      },
-      {
-        id: 'industrial',
-        name: 'Industrial Property',
-        description: 'Manufacturing facilities and industrial properties',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['propertyType', 'location', 'rentalIncome'],
-      },
-    ],
-  },
-  STOCKS: {
-    id: 'stocks',
-    name: 'Stocks & Securities',
-    description: 'Shares, bonds, and other securities',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'individual_stocks',
-        name: 'Individual Stocks',
-        description: 'Shares in individual companies',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['ticker', 'shares', 'dividendYield'],
-      },
-      {
-        id: 'mutual_funds',
-        name: 'Mutual Funds',
-        description: 'Professionally managed investment funds',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['ticker', 'shares', 'dividendYield'],
-      },
-      {
-        id: 'etfs',
-        name: 'Exchange-Traded Funds',
-        description: 'Funds that trade like stocks',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['ticker', 'shares', 'dividendYield'],
-      },
-      {
-        id: 'bonds',
-        name: 'Bonds',
-        description: 'Government and corporate bonds',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['ticker', 'shares', 'dividendYield'],
-      },
-      {
-        id: 'index_funds',
-        name: 'Index Funds',
-        description: 'Funds that track market indices',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['ticker', 'shares', 'dividendYield'],
-      },
-      {
-        id: 'retirement_401k',
-        name: '401(k) Retirement Account',
-        description: 'Employer-sponsored retirement savings plan',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['employerMatch', 'vestingSchedule'],
-      },
-      {
-        id: 'retirement_ira',
-        name: 'IRA (Individual Retirement Account)',
-        description: 'Individual retirement account (traditional or Roth)',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['iraType', 'contributionLimit'],
-      },
-      {
-        id: 'retirement_other',
-        name: 'Other Retirement Accounts',
-        description: 'Other retirement investment accounts (403b, SEP-IRA, etc.)',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['accountType'],
-      },
-      {
-        id: '529_education',
-        name: '529 Education Savings Plan',
-        description: 'Tax-advantaged savings plan for education expenses',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['accountType'],
-      },
-    ],
-  },
-  CRYPTO: {
-    id: 'crypto',
-    name: 'Cryptocurrency',
-    description: 'Digital currencies and crypto assets',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'bitcoin',
-        name: 'Bitcoin',
-        description: 'Bitcoin cryptocurrency',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['coinSymbol', 'quantity', 'stakingRewards'],
-      },
-      {
-        id: 'ethereum',
-        name: 'Ethereum',
-        description: 'Ethereum cryptocurrency',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['coinSymbol', 'quantity', 'stakingRewards'],
-      },
-      {
-        id: 'altcoins',
-        name: 'Alternative Coins',
-        description: 'Other cryptocurrencies besides Bitcoin',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['coinSymbol', 'quantity', 'stakingRewards'],
-      },
-      {
-        id: 'stablecoins',
-        name: 'Stablecoins',
-        description: 'Cryptocurrencies pegged to stable assets',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['coinSymbol', 'quantity'],
-      },
-      {
-        id: 'defi_tokens',
-        name: 'DeFi Tokens',
-        description: 'Decentralized Finance protocol tokens',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['coinSymbol', 'quantity', 'stakingRewards'],
-      },
-    ],
-  },
-  DEBTS: {
-    id: 'debts',
-    name: 'Debts & Receivables',
-    description: 'Money owed to you by others',
-    zakatRate: ZAKAT_RATES.STANDARD_RATE,
-    zakatEligible: true,
-    defaultZakatEligible: true,
-    nisabApplicable: true,
-    subCategories: [
-      {
-        id: 'accounts_receivable',
-        name: 'Accounts Receivable',
-        description: 'Money owed by customers for goods or services',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['debtor', 'dueDate'],
-      },
-      {
-        id: 'personal_loans_given',
-        name: 'Personal Loans Given',
-        description: 'Money lent to individuals',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: [
-          'debtor',
-          'dueDate',
-          'interestRate',
-          'repaymentSchedule',
-        ],
-      },
-      {
-        id: 'business_loans_given',
-        name: 'Business Loans Given',
-        description: 'Money lent to businesses',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: [
-          'debtor',
-          'dueDate',
-          'interestRate',
-          'repaymentSchedule',
-        ],
-      },
-      {
-        id: 'promissory_notes',
-        name: 'Promissory Notes',
-        description: 'Written promises to pay specific amounts',
-        zakatRate: ZAKAT_RATES.STANDARD_RATE,
-        zakatEligible: true,
-        specificFields: ['debtor', 'dueDate', 'interestRate'],
-      },
-    ],
-  },
-  EXPENSES: {
-    id: 'expenses',
-    name: 'Deductible Expenses',
-    description: 'Expenses that can be deducted from zakat calculation',
-    zakatRate: 0, // Expenses reduce zakat, they don't generate it
-    zakatEligible: false,
-    defaultZakatEligible: false,
-    nisabApplicable: false,
-    subCategories: [
-      {
-        id: 'debts_owed',
-        name: 'Outstanding Debts',
-        description: 'Debts you owe to others (credit cards, loans, mortgages)',
-        zakatRate: 0,
-        zakatEligible: false,
-        specificFields: ['creditor', 'dueDate', 'interestRate'],
-      },
-      {
-        id: 'essential_needs',
-        name: 'Essential Living Expenses',
-        description: 'Basic living expenses for the year (housing, food, clothing)',
-        zakatRate: 0,
-        zakatEligible: false,
-        specificFields: ['expenseType', 'frequency'],
-      },
-      {
-        id: 'family_obligations',
-        name: 'Family Financial Obligations',
-        description: 'Financial support for dependents and family members',
-        zakatRate: 0,
-        zakatEligible: false,
-        specificFields: ['dependentCount', 'supportType'],
-      },
-      {
-        id: 'business_liabilities',
-        name: 'Business Liabilities',
-        description: 'Business-related debts and obligations',
-        zakatRate: 0,
-        zakatEligible: false,
-        specificFields: ['businessType', 'liabilityType'],
-      },
-    ],
-  },
-} as const;
-
-/**
- * Zakat calculation methodologies with comprehensive methodology information.
- * 
- * This constant provides detailed information for four main zakat calculation approaches:
- * - Standard Method (AAOIFI-compliant): Modern international consensus approach
- * - Hanafi School Method: Silver-based nisab with comprehensive business inclusion
- * - Shafi'i School Method: Detailed categorization with dual nisab approach
- * - Custom Method: User-defined parameters for unique circumstances
- * 
- * Each methodology includes:
- * @property {string} id - Unique identifier for the methodology
- * @property {string} name - Human-readable name of the methodology
- * @property {string} description - Brief description of the methodology's approach
- * @property {string} nisabBasis - Nisab calculation approach (silver/dual_minimum/configurable)
- * @property {string} businessAssetTreatment - How business assets are handled (market_value/comprehensive/categorized/configurable)
- * @property {string} debtDeduction - Debt deduction approach (immediate/comprehensive/conservative/configurable)
- * @property {string[]} scholarlyBasis - Scholarly sources and jurisprudential foundations
- * @property {string[]} regions - Geographic regions where the methodology is commonly used
- * @property {number} zakatRate - Standard zakat rate percentage (typically 2.5)
- * @property {('lunar'|'solar')[]} calendarSupport - Supported calendar systems
- * @property {boolean} [customRules] - Whether the method supports custom rule configuration
- * @property {string[]} suitableFor - Target user groups and use cases
- * @property {string[]} pros - Advantages and benefits of using this methodology
- * @property {string[]} cons - Limitations and considerations
- * @property {string} explanation - Comprehensive explanation of the methodology's principles and application
- * 
- * This structure supports educational content display, methodology comparison,
- * regional recommendations, and informed user decision-making for zakat calculations.
- * 
- * @see {@link MethodologyInfo} for the TypeScript interface definition
- * @see {@link METHODOLOGY_EDUCATION} for additional educational content
- * @see {@link REGIONAL_METHODOLOGY_MAP} for regional recommendations
- */
-export const ZAKAT_METHODS = {
-  STANDARD: {
-    id: 'standard',
-    name: 'Standard Method (AAOIFI)',
-    description: 'Internationally recognized dual nisab method',
-    nisabBasis: 'dual_minimum',
-    businessAssetTreatment: 'market_value',
-    debtDeduction: 'immediate',
-    scholarlyBasis: ['AAOIFI FAS 9', 'Contemporary consensus'],
-    regions: ['International', 'Gulf States', 'Western countries'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: false,
-    suitableFor: [
-      'Muslims in diverse global locations',
-      'Those seeking internationally recognized standards',
-      'Users preferring simplified calculations',
-      'Multi-national Islamic finance institutions'
-    ],
-    pros: [
-      'Modern consensus approach',
-      'Internationally recognized standards',
-      'Simplified calculation process',
-      'Good for diverse geographic regions',
-      'Supported by contemporary Islamic finance institutions',
-      'Flexible nisab calculation based on market conditions'
-    ],
-    cons: [
-      'May not align with specific regional traditions',
-      'Simplified approach may not capture all nuances',
-      'Requires trust in contemporary scholarly consensus',
-      'Less historical precedent than traditional schools'
-    ],
-    explanation: 'The Standard method represents a modern consensus approach developed for contemporary global Muslim communities. It incorporates guidelines from Islamic finance institutions like AAOIFI and uses a dual minimum nisab approach, selecting the lower threshold between gold and silver to ensure accessibility while maintaining religious compliance.'
-  },
-  HANAFI: {
-    id: 'hanafi',
-    name: 'Hanafi School Method',
-    description: 'Silver-based nisab with comprehensive business inclusion',
-    nisabBasis: 'silver',
-    businessAssetTreatment: 'comprehensive',
-    debtDeduction: 'comprehensive',
-    scholarlyBasis: ['Al-Hidayah', 'Classical Hanafi texts'],
-    regions: ['Turkey', 'Central Asia', 'Indian subcontinent'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: false,
-    suitableFor: [
-      'Muslims following Hanafi jurisprudence',
-      'Residents of Turkey, Central Asia, and South Asia',
-      'Business owners with diverse asset portfolios',
-      'Those preferring comprehensive wealth assessment'
-    ],
-    pros: [
-      'Lower nisab threshold ensures broader zakat eligibility',
-      'Comprehensive business asset inclusion',
-      'Flexible debt deduction approach',
-      'Well-established scholarly precedent',
-      'Thorough wealth assessment methodology',
-      'Accommodates complex business structures'
-    ],
-    cons: [
-      'May result in higher zakat amounts for some individuals',
-      'Requires detailed business asset evaluation',
-      'Complex debt assessment needed',
-      'More time-intensive calculation process'
-    ],
-    explanation: 'The Hanafi method, based on the jurisprudence of Imam Abu Hanifa (699-767 CE), uses exclusively silver-based nisab thresholds and emphasizes comprehensive inclusion of business assets. This approach ensures broader zakat eligibility and thorough wealth assessment, making it particularly suitable for business owners and those with diverse asset portfolios.'
-  },
-  SHAFII: {
-    id: 'shafii',
-    name: 'Shafi\'i School Method',
-    description: 'Detailed categorization with dual nisab',
-    nisabBasis: 'dual_minimum',
-    businessAssetTreatment: 'categorized',
-    debtDeduction: 'conservative',
-    scholarlyBasis: ['Al-Majmu\'', 'Shafi\'i jurisprudence'],
-    regions: ['Southeast Asia', 'East Africa', 'Parts of Middle East'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: false,
-    suitableFor: [
-      'Muslims following Shafi\'i jurisprudence',
-      'Residents of Southeast Asia and East Africa',
-      'Those preferring detailed asset categorization',
-      'Users seeking methodical and systematic approach'
-    ],
-    pros: [
-      'Balanced nisab calculation approach',
-      'Detailed and precise asset categorization',
-      'Conservative debt treatment provides certainty',
-      'Strong methodological framework',
-      'Well-suited for diverse asset types',
-      'Systematic approach reduces calculation errors'
-    ],
-    cons: [
-      'More complex asset categorization required',
-      'Conservative debt approach may limit deductions',
-      'Requires good understanding of different asset types',
-      'May be less accessible for simple portfolios'
-    ],
-    explanation: 'The Shafi\'i method, founded by Imam al-Shafi\'i (767-820 CE), emphasizes systematic methodology and detailed asset categorization. It uses a dual minimum nisab approach while requiring precise classification of business assets and conservative debt treatment, making it ideal for those seeking methodical and thorough zakat calculations.'
-  },
-  MALIKI: {
-    id: 'maliki',
-    name: 'Maliki School Method',
-    description: 'Community-focused approach with regional adaptation',
-    nisabBasis: 'dual_flexible',
-    businessAssetTreatment: 'comprehensive',
-    debtDeduction: 'community_based',
-    scholarlyBasis: ['Al-Mudawwana', 'Bidayat al-Mujtahid', 'Maliki jurisprudence'],
-    regions: ['North Africa', 'West Africa', 'Sudan', 'Parts of Arabia'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: false,
-    suitableFor: [
-      'Muslims following Maliki jurisprudence',
-      'Residents of North and West Africa',
-      'Agricultural communities',
-      'Those preferring community-centric approaches'
-    ],
-    pros: [
-      'Adapts to local economic conditions',
-      'Strong agricultural asset handling',
-      'Community-centric approach',
-      'Flexible implementation based on regional needs',
-      'Considers broader economic context',
-      'Comprehensive trade goods treatment'
-    ],
-    cons: [
-      'Requires regional economic data',
-      'Complex adjustment mechanisms',
-      'Less standardized across regions',
-      'May be difficult to implement uniformly'
-    ],
-    explanation: 'The Maliki method, based on the jurisprudence of Imam Malik (711-795 CE), emphasizes community benefit and practical application. It allows for nisab adjustments based on regional economic conditions and provides detailed rules for agricultural zakat, making it particularly suitable for agricultural communities and regions with varying economic conditions.'
-  },
-  HANBALI: {
-    id: 'hanbali',
-    name: 'Hanbali School Method',
-    description: 'Conservative gold-based approach with textual emphasis',
-    nisabBasis: 'gold',
-    businessAssetTreatment: 'categorized',
-    debtDeduction: 'conservative',
-    scholarlyBasis: ['Al-Mughni', 'Hanbali classical texts', 'Ibn Taymiyyah works'],
-    regions: ['Saudi Arabia', 'Qatar', 'Parts of Gulf States'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: false,
-    suitableFor: [
-      'Muslims following Hanbali jurisprudence',
-      'Residents of Saudi Arabia and Gulf states',
-      'Those preferring traditional textual approaches',
-      'Users seeking conservative calculations'
-    ],
-    pros: [
-      'Clear precedential basis from Quran and Hadith',
-      'Consistent with traditional interpretations',
-      'Simplified calculation logic',
-      'Stable gold-based reference',
-      'Conservative approach ensures compliance',
-      'Well-established scholarly precedent'
-    ],
-    cons: [
-      'May exclude lower-income individuals due to higher thresholds',
-      'Less adaptive to modern financial instruments',
-      'Limited flexibility for contemporary assets',
-      'May not suit all economic contexts'
-    ],
-    explanation: 'The Hanbali method, founded by Imam Ahmad ibn Hanbal (780-855 CE), emphasizes textual adherence and conservative calculations. It prefers gold-based nisab calculations and takes a conservative approach to debt deductions, making it suitable for those seeking traditional, scripture-based zakat calculations with clear precedential foundations.'
-  },
-  CUSTOM: {
-    id: 'custom',
-    name: 'Custom Method',
-    description: 'User-defined calculation parameters',
-    nisabBasis: 'configurable',
-    businessAssetTreatment: 'configurable',
-    debtDeduction: 'configurable',
-    scholarlyBasis: ['User consultation recommended'],
-    regions: ['User-specific'],
-    zakatRate: 2.5,
-    calendarSupport: ['lunar', 'solar'],
-    customRules: true,
-    suitableFor: [
-      'Muslims with unique circumstances requiring specialized calculations',
-      'Those following specific regional practices not covered by standard methods',
-      'Users with access to qualified Islamic scholars for consultation',
-      'Communities with established local zakat practices'
-    ],
-    pros: [
-      'Maximum flexibility for unique situations',
-      'Can accommodate specific regional practices',
-      'Allows for scholarly consultation integration',
-      'Adaptable to changing circumstances',
-      'Supports specialized business models',
-      'Enables compliance with local Islamic authorities'
-    ],
-    cons: [
-      'Requires qualified scholarly guidance',
-      'May lack standardization benefits',
-      'Responsibility for correctness lies with user',
-      'May be complex to implement properly',
-      'Potential for inconsistent applications',
-      'Requires ongoing scholarly oversight'
-    ],
-    explanation: 'The Custom method provides maximum flexibility for users with unique circumstances or specific regional requirements. It allows configuration of all calculation parameters based on scholarly consultation or established local practices. This method requires careful implementation with qualified Islamic scholarship to ensure religious compliance while meeting specific needs.'
-  }
-} as const;
+// @deprecated Use ASSET_CATEGORIES from islamicConstants.ts
 
 // Calendar types
 export const CALENDAR_TYPES = {
