+++ shared/src/types/index.ts
@@ -1,7 +1,7 @@
 /**
  * Re-export all types from subdirectories
  */
-export * from './auditTrail';
-export * from './hawl';
-export * from './nisabYearRecord';
-export * from './tracking';
+export * from './auditTrail.js';
+export * from './hawl.js';
+export * from './nisabYearRecord.js';
+export * from './tracking.js';
